src/test/javascript/spec/app/entities/immatriculation/immatriculation.service.spec.ts
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import * as moment from 'moment';
import { DATE_TIME_FORMAT } from 'app/shared/constants/input.constants';
import { ImmatriculationService } from 'app/entities/immatriculation/immatriculation.service';
import { IImmatriculation, Immatriculation } from 'app/shared/model/immatriculation.model';

describe('Service Tests', () => {
  describe('Immatriculation Service', () => {
    let injector: TestBed;
    let service: ImmatriculationService;
    let httpMock: HttpTestingController;
    let elemDefault: IImmatriculation;
    let expectedResult: IImmatriculation | IImmatriculation[] | boolean | null;
    let currentDate: moment.Moment;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule]
      });
      expectedResult = null;
      injector = getTestBed();
      service = injector.get(ImmatriculationService);
      httpMock = injector.get(HttpTestingController);
      currentDate = moment();

      elemDefault = new Immatriculation(0, 'AAAAAAA', currentDate);
    });

    describe('Service methods', () => {
      it('should find an element', () => {
        const returnedFromService = Object.assign(
          {
            dateImmatriculation: currentDate.format(DATE_TIME_FORMAT)
          },
          elemDefault
        );

        service.find(123).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(elemDefault);
      });

      it('should create a Immatriculation', () => {
        const returnedFromService = Object.assign(
          {
            id: 0,
            dateImmatriculation: currentDate.format(DATE_TIME_FORMAT)
          },
          elemDefault
        );

        const expected = Object.assign(
          {
            dateImmatriculation: currentDate
          },
          returnedFromService
        );

        service.create(new Immatriculation()).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'POST' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should update a Immatriculation', () => {
        const returnedFromService = Object.assign(
          {
            numero: 'BBBBBB',
            dateImmatriculation: currentDate.format(DATE_TIME_FORMAT)
          },
          elemDefault
        );

        const expected = Object.assign(
          {
            dateImmatriculation: currentDate
          },
          returnedFromService
        );

        service.update(expected).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'PUT' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should return a list of Immatriculation', () => {
        const returnedFromService = Object.assign(
          {
            numero: 'BBBBBB',
            dateImmatriculation: currentDate.format(DATE_TIME_FORMAT)
          },
          elemDefault
        );

        const expected = Object.assign(
          {
            dateImmatriculation: currentDate
          },
          returnedFromService
        );

        service.query().subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush([returnedFromService]);
        httpMock.verify();
        expect(expectedResult).toContainEqual(expected);
      });

      it('should delete a Immatriculation', () => {
        service.delete(123).subscribe(resp => (expectedResult = resp.ok));

        const req = httpMock.expectOne({ method: 'DELETE' });
        req.flush({ status: 200 });
        expect(expectedResult);
      });
    });

    afterEach(() => {
      httpMock.verify();
    });
  });
});
